Add launcher for "updates"
This patch adds an extra, invisible, .desktop file that opens
the store in the "updates" page directly.

This is needed because in the new Refresh App Awareness, we
want to allow the user to open the snap-center in the "updates"
page directly from a "pending update" notification, but since
it is `snapd-desktop-integration` the application that opens
it, and it is inside another snap, the only way that it has
to launch the `snap-center` is using the `desktop-launch` DBus
interface. This interface, unfortunately, doesn't allow to
pass parameters, only allows to pass the name of a `.desktop`
file to launch, so this workaround is needed.